validation fix

+++ src/app/models/dive.py
@@ -21,7 +21,7 @@ class Dive(Base):
 
     max_depth: Mapped[float | None] = mapped_column(Float, default=None)
     avg_depth: Mapped[float | None] = mapped_column(Float, default=None)
-    bottom_temperature: Mapped[int | None] = mapped_column(Integer, default=None)
+    bottom_temperature: Mapped[float | None] = mapped_column(Float, default=None)
     visibility: Mapped[int | None] = mapped_column(Integer, default=None)
     trip_id: Mapped[int | None] = mapped_column(
         ForeignKey("trip.id", ondelete="SET NULL"), default=None, index=True
