+++ src/app/schemas/dive.py
@@ -14,7 +14,7 @@ class DiveBase(BaseModel):
 
     max_depth: Annotated[float | None, Field(default=None)]
     avg_depth: Annotated[float | None, Field(default=None)]
-    bottom_temperature: Annotated[int | None, Field(default=None)]
+    bottom_temperature: Annotated[float | None, Field(default=None)]
     visibility: Annotated[int | None, Field(default=None, description="Underwater visibility in meters")]
     trip_id: Annotated[int | None, Field(default=None, description="ID of the trip this dive belongs to")]
     dive_site_id: Annotated[int | None, Field(default=None, description="ID of the dive site this dive was made at")]
@@ -66,7 +66,7 @@ class DiveUpdate(BaseModel):
     ]
     max_depth: Annotated[float | None, Field(default=None)]
     avg_depth: Annotated[float | None, Field(default=None)]
-    bottom_temperature: Annotated[int | None, Field(default=None)]
+    bottom_temperature: Annotated[float | None, Field(default=None)]
     visibility: Annotated[int | None, Field(default=None, description="Underwater visibility in meters")]
     trip_id: Annotated[int | None, Field(default=None, description="ID of the trip this dive belongs to")]
     dive_site_id: Annotated[int | None, Field(default=None, description="ID of the dive site this dive was made at")]
